src/EMRNext.Core/Infrastructure/Interoperability/DataInteroperabilityService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.Extensions.Logging;
using Hl7.Fhir.Model;
using Hl7.Fhir.Serialization;
using Newtonsoft.Json;
using Newtonsoft.Json.Linq;

namespace EMRNext.Core.Infrastructure.Interoperability
{
    public class DataInteroperabilityService
    {
        // Data Source Types
        public enum DataSourceType
        {
            FHIR,
            HL7V2,
            DICOM,
            Custom
        }

        // Transformation Strategy
        public enum TransformationStrategy
        {
            Direct,
            Normalized,
            Enriched
        }

        // Mapping Configuration
        public class DataMappingConfiguration
        {
            public string SourceSystem { get; set; }
            public string TargetSystem { get; set; }
            public DataSourceType SourceType { get; set; }
            public DataSourceType TargetType { get; set; }
            public TransformationStrategy Strategy { get; set; }
            public Dictionary<string, string> FieldMappings { get; set; }
            public List<string> RequiredFields { get; set; }
        }

        // Transformation Result
        public class TransformationResult
        {
            public bool IsSuccessful { get; set; }
            public object TransformedData { get; set; }
            public List<string> ValidationErrors { get; set; }
            public DataSourceType SourceType { get; set; }
            public DataSourceType TargetType { get; set; }
        }

        // Data Validation Report
        public class DataValidationReport
        {
            public bool IsValid { get; set; }
            public List<ValidationIssue> Issues { get; set; }
        }

        // Validation Issue
        public class ValidationIssue
        {
            public string Field { get; set; }
            public string Message { get; set; }
            public ValidationSeverity Severity { get; set; }
        }

        // Validation Severity
        public enum ValidationSeverity
        {
            Information,
            Warning,
            Error
        }

        private readonly ILogger<DataInteroperabilityService> _logger;

        public DataInteroperabilityService(
            ILogger<DataInteroperabilityService> logger)
        {
            _logger = logger;
        }

        // Transform Data Between Different Standards
        public async Task<TransformationResult> TransformDataAsync(
            object sourceData, 
            DataMappingConfiguration mappingConfig)
        {
            try
            {
                // Validate input data
                var validationReport = ValidateSourceData(sourceData, mappingConfig);
                if (!validationReport.IsValid)
                {
                    return new TransformationResult
                    {
                        IsSuccessful = false,
                        ValidationErrors = validationReport.Issues
                            .Select(i => i.Message)
                            .ToList(),
                        SourceType = mappingConfig.SourceType,
                        TargetType = mappingConfig.TargetType
                    };
                }

                // Transform based on strategy
                object transformedData = mappingConfig.Strategy switch
                {
                    TransformationStrategy.Direct => 
                        await DirectTransformAsync(sourceData, mappingConfig),
                    TransformationStrategy.Normalized => 
                        await NormalizedTransformAsync(sourceData, mappingConfig),
                    TransformationStrategy.Enriched => 
                        await EnrichedTransformAsync(sourceData, mappingConfig),
                    _ => throw new NotSupportedException(
                        $"Transformation strategy not supported: {mappingConfig.Strategy}")
                };

                return new TransformationResult
                {
                    IsSuccessful = true,
                    TransformedData = transformedData,
                    SourceType = mappingConfig.SourceType,
                    TargetType = mappingConfig.TargetType
                };
            }
            catch (Exception ex)
            {
                _logger.LogError(ex, "Data transformation failed");
                return new TransformationResult
                {
                    IsSuccessful = false,
                    ValidationErrors = new List<string> { ex.Message },
                    SourceType = mappingConfig.SourceType,
                    TargetType = mappingConfig.TargetType
                };
            }
        }

        // Direct Data Transformation
        private async Task<object> DirectTransformAsync(
            object sourceData, 
            DataMappingConfiguration mappingConfig)
        {
            // Simple field mapping without complex transformations
            var sourceDict = ConvertToDictionary(sourceData);
            var targetDict = new Dictionary<string, object>();

            foreach (var mapping in mappingConfig.FieldMappings)
            {
                if (sourceDict.ContainsKey(mapping.Key))
                {
                    targetDict[mapping.Value] = sourceDict[mapping.Key];
                }
            }

            return await Task.FromResult(targetDict);
        }

        // Normalized Data Transformation
        private async Task<object> NormalizedTransformAsync(
            object sourceData, 
            DataMappingConfiguration mappingConfig)
        {
            // More complex transformation with data normalization
            var sourceDict = ConvertToFhirResource(sourceData);
            var targetResource = new Patient();

            // Example FHIR mapping logic
            if (sourceDict is Patient sourceFhirPatient)
            {
                targetResource.Name = sourceFhirPatient.Name;
                targetResource.Identifier = sourceFhirPatient.Identifier;
                targetResource.Gender = sourceFhirPatient.Gender;
                
                // Additional normalization logic
                NormalizePatientData(targetResource);
            }

            return await Task.FromResult(targetResource);
        }

        // Enriched Data Transformation
        private async Task<object> EnrichedTransformAsync(
            object sourceData, 
            DataMappingConfiguration mappingConfig)
        {
            // Most complex transformation with data enrichment
            var baseTransformation = await NormalizedTransformAsync(sourceData, mappingConfig);
            
            // Enrich with additional data sources or external services
            var enrichedData = await EnrichDataAsync(baseTransformation);

            return enrichedData;
        }

        // Data Enrichment
        private async Task<object> EnrichDataAsync(object baseData)
        {
            // Implement data enrichment logic
            // Could involve calling external APIs, 
            // adding missing information, etc.
            return await Task.FromResult(baseData);
        }

        // Validate Source Data
        private DataValidationReport ValidateSourceData(
            object sourceData, 
            DataMappingConfiguration mappingConfig)
        {
            var issues = new List<ValidationIssue>();

            // Check required fields
            var sourceDict = ConvertToDictionary(sourceData);
            foreach (var requiredField in mappingConfig.RequiredFields)
            {
                if (!sourceDict.ContainsKey(requiredField))
                {
                    issues.Add(new ValidationIssue
                    {
                        Field = requiredField,
                        Message = $"Required field {requiredField} is missing",
                        Severity = ValidationSeverity.Error
                    });
                }
            }

            return new DataValidationReport
            {
                IsValid = issues.Count == 0,
                Issues = issues
            };
        }

        // Convert Object to Dictionary
        private Dictionary<string, object> ConvertToDictionary(object data)
        {
            if (data is JObject jObject)
            {
                return jObject.ToObject<Dictionary<string, object>>();
            }

            var json = JsonConvert.SerializeObject(data);
            return JsonConvert.DeserializeObject<Dictionary<string, object>>(json);
        }

        // Convert to FHIR Resource
        private Resource ConvertToFhirResource(object data)
        {
            if (data is Resource fhirResource)
                return fhirResource;

            var parser = new FhirJsonParser();
            var json = JsonConvert.SerializeObject(data);
            return parser.Parse<Resource>(json);
        }

        // Normalize Patient Data
        private void NormalizePatientData(Patient patient)
        {
            // Standardize name formatting
            if (patient.Name != null)
            {
                foreach (var name in patient.Name)
                {
                    name.Family = NormalizeString(name.Family);
                    name.Given = name.Given?.Select(NormalizeString).ToList();
                }
            }

            // Standardize identifier
            if (patient.Identifier != null)
            {
                foreach (var identifier in patient.Identifier)
                {
                    identifier.Value = NormalizeIdentifier(identifier.Value);
                }
            }
        }

        // Normalize String
        private string NormalizeString(string input)
        {
            if (string.IsNullOrWhiteSpace(input))
                return input;

            return input.Trim().ToUpperInvariant();
        }

        // Normalize Identifier
        private string NormalizeIdentifier(string identifier)
        {
            if (string.IsNullOrWhiteSpace(identifier))
                return identifier;

            // Remove spaces, dashes, and convert to uppercase
            return new string(
                identifier
                    .Replace(" ", "")
                    .Replace("-", "")
                    .ToUpperInvariant()
                    .Where(char.IsLetterOrDigit)
                    .ToArray()
            );
        }
    }
}
